remap based the objectId of the largest database
When the target database is smaller than the child we can have
object_id clashes since adding the maxId of the smaller database to
a value from the larger database is not guaranteed to be outside of
that database's used object_id range.

Using the largest value from whichever database is the largest should
avoid this issue.