handle conflicts with character object IDs
Remapping character object IDs results in loss of ownership of
Thralls. Thrall ownership is likely stored as a binary property of
the Thrall object, which we don't really want to mess with at this
point.

To still avoid ownership issues we now prefer to remap the
non-character object ID in the target database, rather than remapping
the character in the source database like we did before.

The case where two character object IDs conflict currently is not
handled: the conflicting characters are just ignored during the copy
resulting in the character in the source database getting dropped.

In the future we want to copy those anyway, but warn the user about
Thrall ownership issues on this character so they can make an
informed decision on whether to go ahead or not.