remap static_buildables to avoid conflicts
Static buildables are a relatively new table added by Funcom, the
table appears to handle "buildables" that were placed by the devs,
like the various forges in dungeons.

However this table appears to only get populated once a player has
interacted with one of these buildables, meaning the IDs are not
static, and as such susceptible to cross-savegame conflicts.