remove duplicate objects prior to merging
When merging two savegames with the same parent they will likely
contain many objects that are the same, and as such, are in exactly
the same location in the game world.

Doing a normal merge would result in all of these getting duplicated,
so we attempt to find these objects and remove them from the target
prior to merging.

We ignore the last set or coordinates present in the game database
(rx, ry, rz, rw) as these appear to often be different, even for
objects that are the same. Presumably these might be related to NPC
positioning, or "stance"